% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_muller}
\alias{cc_muller}
\title{Replication Data for: The Temporal Focus of Campaign Communication (2020 JOP)}
\format{
A data frame with 7 variables:
\describe{
  \item{text}{Content of the text.}
  \item{sentence_id}{Unique identifier for each sentence.}
  \item{countryname}{Country's name.}
  \item{party}{Associated political party of the text.}
  \item{date}{Date of the record.}
  \item{class}{Type or classification.}
  \item{class_pro_retro}{Classification as either 'Prospective' or 'Retrospective'.}
}
}
\source{
Data provided by the author
and \url{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/7NP2XH}
}
\usage{
data("cc_muller")
}
\description{
The replication data is sourced from "The Temporal Focus of Campaign
Communication," authored by Stefan Müller, and was published in the Journal of Politics in 2022.
This study primarily delves into the temporal emphasis of party manifestos. The dataset
encompasses 5,858 annotated data entries from countries such as the United Kingdom,
Ireland, Canada, Australia, New Zealand, and the United States. Its central objective
is to compute the percentage of sentences or quasi-sentences referring to the past, present,
or future. This differentiation is made based on two categories: "Prospective" and "Retrospective".
The paper can be accessed at \url{https://www.journals.uchicago.edu/doi/10.1086/715165}.
}
\examples{
\dontrun{
data(cc_muller)
head(cc_muller)
}
}
