% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_and_gc}
\alias{check_and_gc}
\title{Perform Garbage Collection Based on Condition}
\usage{
check_and_gc(gc.active)
}
\arguments{
\item{gc.active}{A logical value indicating whether or not to
activate garbage collection.}
}
\value{
A message indicating that garbage collection was performed
    if `gc.active` was `TRUE`. Otherwise, no action is taken or message is displayed.
}
\description{
This function checks the value of `gc.active` to determine whether
or not to perform garbage collection. If `gc.active` is `TRUE`,
the function will perform garbage collection and then send a
message indicating the completion of this process.
}
\examples{
check_and_gc(TRUE)  # Performs garbage collection and displays a message
check_and_gc(FALSE) # No action or message

}
\keyword{internal}
