% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_batch_size}
\alias{check_batch_size}
\title{Check the specified batch size}
\usage{
check_batch_size(batch_size)
}
\arguments{
\item{batch_size}{Integer. The batch size to be checked.}
}
\description{
Validates if the given batch size is a positive integer.
}
\keyword{internal}
