% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_flair_installed}
\alias{check_flair_installed}
\title{Check Flair}
\usage{
check_flair_installed(...)
}
\value{
Logical. `TRUE` if Flair is installed, otherwise `FALSE`.
}
\description{
Determines if the Flair Python module is available in the current Python
environment.
}
\keyword{internal}
