% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_language_supported}
\alias{check_language_supported}
\title{Check the given language against supported languages}
\usage{
check_language_supported(language, supported_lan_models)
}
\arguments{
\item{language}{The language to check.}

\item{supported_lan_models}{A vector of supported languages.}
}
\value{
This function does not return anything, but stops execution if the check fails.
}
\description{
This function checks whether a provided language is supported. If it's not,
it stops the execution and returns a message indicating the supported languages.
}
\examples{
# Assuming 'en' is a supported language and 'abc' is not:
check_language_supported("en", c("en", "de", "fr"))
# check_language_supported("abc", c("en", "de", "fr")) # will stop execution

}
