% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_prerequisites}
\alias{check_prerequisites}
\title{Check Environment Pre-requisites}
\usage{
check_prerequisites(...)
}
\arguments{
\item{...}{passing additional arguments.}
}
\value{
A message detailing any missing pre-requisites.
}
\description{
This function checks if Python is installed, if the flair module
is available in Python,
and if there's an active internet connection.
}
