% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_python_environment}
\alias{check_python_environment}
\title{Check for Available Python Environment}
\usage{
check_python_environment(...)
}
\arguments{
\item{...}{Additional arguments passed to next.}
}
\value{
Logical indicating if a Python environment is available.
}
\description{
This function checks if a Python environment is available
and prints the path to the Python executable if it is.
}
