% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_python_installed}
\alias{check_python_installed}
\title{Check if Python is Installed}
\usage{
check_python_installed()
}
\value{
Logical. `TRUE` if Python is installed, `FALSE` otherwise. Additionally, if installed, the path to the Python installation is printed.
}
\description{
This function checks if Python is installed on the system.
}
\examples{
check_python_installed()

}
