% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_show.text_id}
\alias{check_show.text_id}
\title{Check the `show.text_id` Parameter}
\usage{
check_show.text_id(show.text_id)
}
\arguments{
\item{show.text_id}{Logical. The parameter to be checked.}
}
\description{
Validates if the given `show.text_id` is a logical value.
}
\keyword{internal}
