% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entities.R
\name{check_tagger}
\alias{check_tagger}
\title{Check if Tagger is Valid}
\usage{
check_tagger(tagger)
}
\arguments{
\item{tagger}{A Flair tagger object to check}
}
\value{
Logical indicating if the tagger is valid
}
\description{
Internal function to verify if the provided tagger is valid
and has the required methods.
}
\keyword{internal}
