% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_texts_and_ids}
\alias{check_texts_and_ids}
\title{Check the texts and document IDs}
\usage{
check_texts_and_ids(texts, doc_ids)
}
\arguments{
\item{texts}{List. A list of texts.}

\item{doc_ids}{List. A list of document IDs.}
}
\description{
Validates if the given texts and document IDs are not NULL or empty.
}
\keyword{internal}
