% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_torch_version}
\alias{check_torch_version}
\title{Check the installed version of the `torch` library in Python}
\usage{
check_torch_version()
}
\value{
A list containing a string with the version check result and a boolean
indicating success (TRUE) or failure (FALSE).
}
\description{
This function checks the installed version of the `torch` library
in Python. It returns a list containing a formatted string indicating
success or failure  and a boolean indicating the result.
}
\keyword{internal}
