% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear_flair_cache}
\alias{clear_flair_cache}
\title{Clear Flair Cache}
\usage{
clear_flair_cache(...)
}
\arguments{
\item{...}{The argument passed to next.}
}
\value{
Returns NULL invisibly. Messages are printed indicating whether the cache was found and cleared.
}
\description{
This function clears the cache associated with the Flair Python library.
The cache directory is typically located at "~/.flair".
}
\examples{
\dontrun{
clear_flair_cache()
}
}
