% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_flair_env}
\alias{create_flair_env}
\title{Create and Use a Conda Environment for Flair
s
This function creates a new conda environment specifically for Flair, restarts
the R session, and installs Flair using pip.}
\usage{
create_flair_env(env_name = "flair_env", python_ver = "3.8")
}
\arguments{
\item{env_name}{Name of the conda environment. Defaults to "flair_env".}

\item{python_ver}{Python version to be used in the conda environment. Defaults to "3.8".}
}
\value{
The path to the Python executable in the new environment.
}
\description{
Create and Use a Conda Environment for Flair
s
This function creates a new conda environment specifically for Flair, restarts
the R session, and installs Flair using pip.
}
