% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{de_immigration}
\alias{de_immigration}
\title{German Bundestag Immigration Debate Data}
\format{
A data frame with 16 variables:
\describe{
  \item{date}{Date of the speech, Date type}
  \item{agenda}{Agenda or subject of the speech, character}
  \item{speechnumber}{Unique identifier for each speech, numeric}
  \item{speaker}{Name of the person giving the speech, character}
  \item{party}{Political party of the speaker, character}
  \item{party.facts.id}{ID for the party, usually numeric or character}
  \item{chair}{Person chairing the session, character}
  \item{terms}{Terms or tags associated with the speech, character or list}
  \item{text}{Actual text of the speech, character}
  \item{parliament}{Which Bundestag session, character or numeric}
  \item{iso3country}{ISO3 country code for Germany, character}
  \item{year}{Year when the speech was made, numeric}
  \item{agenda_ID}{Unique identifier for the agenda, usually numeric or character}
  \item{migration_dummy}{Dummy variable related to migration topic, usually numeric (0 or 1)}
  \item{comment_agenda}{Additional comments on the agenda, character}
}
}
\source{
Describe the source of your data here.
}
\usage{
data("de_immigration")
}
\description{
A dataset containing speeches and debates from the German Bundestag on the topic of immigration.
}
\examples{
\dontrun{
data(de_immigration)
head(de_immigration)
}
}
