% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_flair.R
\name{decorated}
\alias{decorated}
\title{Create a \code{decorated} object}
\usage{
decorated(x)
}
\arguments{
\item{x}{A text object or code chunk label}
}
\value{
An object of class \code{decorated}
}
\description{
(The preferred function is \code{\link{decorate}})
}
