% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{On Load Function for Flair Package}
\usage{
.onLoad(libname, pkgname = "flair")
}
\arguments{
\item{libname}{Name of the library (not typically used in the function body).}

\item{pkgname}{Name of the package, defaults to "flair".}
}
\description{
This function is executed when the flair package is loaded.
It prints a message indicating that the package has been loaded and shows its version.
}
