% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{.pkgenv}
\alias{.pkgenv}
\title{Package environment}
\format{
An object of class \code{environment} of length 4.
}
\usage{
.pkgenv
}
\description{
Internal environment for storing package state
}
\keyword{internal}
