% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{embeddings_to_matrix}
\alias{embeddings_to_matrix}
\title{Convert Embeddings to Matrix}
\usage{
embeddings_to_matrix(embeddings, strategy = "average")
}
\arguments{
\item{embeddings}{A three-dimensional array of shape (number_of_texts, number_of_words, embedding_dimension).}

\item{strategy}{A character string specifying the strategy to use. Options are "average", "concatenate",
"max_pooling", and "min_pooling".}
}
\value{
A two-dimensional matrix with the transformed embeddings.
}
\description{
This function takes a three-dimensional array of embeddings and converts them to a two-dimensional matrix
based on the specified strategy.
}
\examples{
\dontrun{
embeddings <- array(runif(10 * 5 * 3), c(10, 5, 3))
result <- embeddings_to_matrix(embeddings, strategy = "average")
}
}
