% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure_tagger_or_language}
\alias{ensure_tagger_or_language}
\title{Ensure Either a Tagger Object or a Language is Specified}
\usage{
ensure_tagger_or_language(tagger = NULL, language = NULL, alternative = NULL)
}
\arguments{
\item{tagger}{A tagger object (default is NULL).}

\item{language}{The language of the texts (default is NULL).}
}
\value{
None. The function will throw a message if neither tagger nor language is specified.
}
\description{
This function checks if either a tagger object or a language is specified and throws an error if neither is provided.
}
\keyword{internal}
