% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{flaiR-package}
\alias{flaiR}
\alias{flaiR-package}
\title{Initialize Python Environment for flaiR NLP Package}
\description{
Initializes Python environment and sets up flair NLP with a visual progress display.
}
\note{
The package automatically detects and uses the appropriate Python installation.
If initialization fails, please ensure you have a compatible Python version installed
and restart your R session.
}
\section{Progress Stages}{

The initialization process is displayed with a progress bar showing three main stages:
\enumerate{
\item Setting up Python environment
\item Installing dependencies
\item Initializing Flair NLP
}
}

\section{Python Requirements}{

\itemize{
\item Python version 3.9-3.12
\item Key dependencies:
\itemize{
\item numpy (1.26.4)
\item scipy (1.12.0)
\item flair (>=0.11.3)
}
}
}

\section{Status Display}{

The package provides visual feedback during initialization:
\preformatted{
[==========>              ] Setting up Python environment...
[====================>    ] Installing dependencies...
[==============================>] Initializing Flair NLP...
Ready!
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://davidycliao.github.io/flaiR}
  \item Report bugs at \url{https://github.com/davidycliao/flaiR/issues}
}

}
\author{
\strong{Maintainer}: Yen-Chieh Liao \email{davidyclaio@gmail.com} [contributor]

Authors:
\itemize{
  \item Stefan Müller \email{stefan.mueller@ucd.ie} [contributor]
  \item Akbik Alan \email{alan.akbik@zalando.de} [contributor]
  \item Blythe Duncan \email{blythe.duncan@zalando.de} [contributor]
  \item Vollgraf Roland \email{vollgraf.roland@zalando.de} [contributor]
}

Other contributors:
\itemize{
  \item Sohini Timbadia [contributor]
}

}
