% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapping_python_flair.R
\name{flair.embeddings.TransformerWordEmbeddings}
\alias{flair.embeddings.TransformerWordEmbeddings}
\title{Create a Flair TransformerWordEmbeddings Object}
\usage{
flair.embeddings.TransformerWordEmbeddings(
  pre_trained_model = "bert-base-uncased"
)
}
\arguments{
\item{pre_trained_model}{A character string specifying the pre-trained model to use.
Defaults to 'bert-base-uncased'.}
}
\value{
A Flair TransformerWordEmbeddings object.
}
\description{
This function interfaces with Python via `reticulate` to create
a `TransformerWordEmbeddings` object using the Flair library.
}
\examples{
\dontrun{
embedding <- flair_data_sentence("bert-base-uncased")
}
}
\references{
Python equivalent:
\preformatted{
from flair.embeddings import TransformerWordEmbeddings
embedding = TransformerWordEmbeddings('bert-base-uncased')
}
}
