% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair.R
\docType{package}
\name{flair}
\alias{flair}
\alias{flair-package}
\title{flair: A package for adding flair to your displayed source code.}
\description{
The flair package provides ...
}
\section{Highlight functions}{


The \code{hlt_*} functions add display formatting to source code.
}

\section{Flair functions}{


The \code{flair_*} functions execute the code and prepare source code to be printed with flair when knitted.
}

