% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_chunk.R
\name{flair_chunk}
\alias{flair_chunk}
\title{Builds a \code{\link{flair_code}} object from a code chunk}
\usage{
flair_chunk(chunk_name)
}
\arguments{
\item{chunk_name}{String that gives the name of the chunk. If left blank, current chunk is used.}
}
\value{
An object of class \code{\link{flair_code}}
}
\description{
This function reads the source code from a given named code chunk; i.e., \code{{r chunk_name, echo = FALSE}}.
}
\details{
When run directly in a source file, \code{flair_chunk()} reads the text of the active file and extracts the relevant string of source code from the chosen chunk.  (Important: this only works in RStudio.)

When run during the \code{knitr::knit()} process, \code{flair_chunk()} pulls the relevant chunk source during \code{knitr::knit_hooks$set("source").}
}
