% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flar_data.R
\name{flair_data}
\alias{flair_data}
\title{Import flair.data Module}
\usage{
flair_data()
}
\value{
A Python module (`flair.data`).
}
\description{
The `flair.data` module provides essential utilities for text data
processing and representation in the Flair library. This function
gives access to various classes and utilities in the `flair.data` module,
most notably:
\itemize{
  \item \strong{Sentence}: Represents a sentence, which is a list of
  Tokens. This class provides various utilities for sentence
  manipulation, such as adding tokens, tagging with pre-trained models,
  and obtaining embeddings.
  \item \strong{Token}: Represents a word or a sub-word unit in a sentence.
  It can carry various annotations such as named entity tags, part-of-speech
  tags, and embeddings. Additionally, the token provides functionalities
  to retrieve or check its annotations.
  \item \strong{Corpus}: Represents a collection of sentences,
  facilitating operations like splitting into train/test/development
   sets and applying transformations. It is particularly useful
   for training and evaluating models on custom datasets.
}
Additionally, the module offers utilities for reading data in the CoNLL
format, a common format for NER, POS tagging, and more. It also contains
the `Dictionary` class for item-index mapping, facilitating the conversion
of text into machine-readable formats. This function provides a bridge
to access these functionalities directly from R.
}
\examples{
\dontrun{
Sentence <- flair_data()$Sentence
Token <- flair_data()$Token
Corpus <- flair_data()$Corpus
}

}
\references{
Python reference:
\preformatted{
from flair.data import Sentence
}
}
