% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flar_data.R
\name{flair_data.Sentence}
\alias{flair_data.Sentence}
\title{Create a Flair Sentence}
\usage{
flair_data.Sentence(sentence_text)
}
\arguments{
\item{sentence_text}{A character string to be converted into a Flair Sentence
object.}
}
\value{
A Flair Sentence object.
}
\description{
Flair is a powerful NLP framework that leverages state-of-the-art
embeddings for various natural language processing tasks.
}
\note{
Ensure the input string is in a language compatible with the intended
Flair model. In R, when processing multiple text, you can use purrr or
the basic R functions lapply and sapply.
}
\examples{
\dontrun{
flair_data.Sentence("The quick brown fox jumps over the lazy dog.")}

}
\references{
Python equivalent: \preformatted{ from flair.data import Sentence
sentence = Sentence("The quick brown fox jumps over the lazy dog.")
}
}
\seealso{
[Flair's GitHub Repository about Senetence object.](https://flairnlp.github.io/docs/tutorial-basics/basic-types)
}
