% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapping_python_flair.R
\name{flair_datasets}
\alias{flair_datasets}
\title{Access the flair_datasets Module from Flair}
\usage{
flair_datasets()
}
\value{
A Python Module(flair.datasets) from Flair, which can be utilized for NLP tasks.
}
\description{
Utilizes the {reticulate} package to import the `flair.datasets`
dataset from Flair's datasets in Python, enabling the use of this dataset in
an R environment.
}
\examples{
\dontrun{
UD_ENGLISH <- flair_datasets()$UD_ENGLISH
corpus <- UD_ENGLISH()$downsample(0.1)
}

}
\references{
Python equivalent:
\preformatted{
from flair.datasets import UD_ENGLISH
corpus = UD_ENGLISH().downsample(0.1)
}
}
\seealso{
\url{https://github.com/flairNLP/flair} for additional information on Flair's capabilities and datasets in NLP.
}
