% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapping_python_flair.R
\name{flair_embeddings.FlairEmbeddings}
\alias{flair_embeddings.FlairEmbeddings}
\title{Flair Embedding Initialization}
\usage{
flair_embeddings.FlairEmbeddings(embeddings_type = "news-forward")
}
\arguments{
\item{embeddings_type}{Character, type of embeddings to initialize.
Options: "news-forward", "news-backward".}
}
\value{
A Flair embeddings object from Python's Flair library.
}
\description{
This function initializes Flair embeddings using Python's Flair
library.
}
\examples{
\dontrun{
flair_embedding_forward <- flair_embeddings.FlairEmbeddings("news-forward")
flair_embedding_backward <- flair_embeddings.FlairEmbeddings("news-backward")
}

}
\references{
FlairEmbeddings from Flair library in Python. Example usage in Python:
\preformatted{
flair_embedding_forward = FlairEmbeddings('news-forward')
flair_embedding_backward = FlairEmbeddings('news-backward')
}
}
