% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_embeddings.R
\name{flair_embeddings.StackedEmbeddings}
\alias{flair_embeddings.StackedEmbeddings}
\title{Initializing a Class for StackedEmbeddings}
\usage{
flair_embeddings.StackedEmbeddings(embeddings_list)
}
\arguments{
\item{embeddings_list}{A list containing Flair embedding instances.}
}
\value{
An instance of the StackedEmbeddings from the flair.embeddings module.
}
\description{
Creates a stacked embedding instance using multiple Flair embeddings.
}
\details{
The function ensures that each embedding provided in the list is a recognized Flair embedding.
If any of the embeddings in the list is not recognized, the function will throw an error.
}
\examples{
\dontrun{
glove_embedding <- flair_embeddings.WordEmbeddings("glove")
fasttext_embedding <- flair_embeddings.WordEmbeddings("fasttext")
stacked_embedding <- flair_embeddings.StackedEmbeddings(list(glove_embedding, fasttext_embedding))
}

}
