% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_lines.R
\name{flair_lines.decorated}
\alias{flair_lines.decorated}
\title{S3 method for \code{\link{decorated}} objects}
\usage{
\method{flair_lines}{decorated}(x, lines)
}
\arguments{
\item{x}{An object of class \code{\link{decorated}}.}

\item{lines}{An integer vector specifying code lines to highlight.}
}
\value{
An object of class \code{\link{decorated}}.
}
\description{
Applies flair to the appropriate line(s) of source code.
}
