% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_lines.R
\name{flair_lines.with_flair}
\alias{flair_lines.with_flair}
\title{S3 method for \code{\link{with_flair}} objects}
\usage{
\method{flair_lines}{with_flair}(x, lines)
}
\arguments{
\item{x}{An object of class \code{\link{with_flair}}.}

\item{lines}{An integer vector specifying code lines to highlight.}
}
\value{
An object of class \code{\link{with_flair}}.
}
\description{
Applies flair to the appropriate line(s) of source code.
}
