% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapping_python_flair.R
\name{flair_models}
\alias{flair_models}
\title{Import the flair.models Python module}
\usage{
flair_models()
}
\value{
A Python module object representing flair.models.
}
\description{
This function utilizes the reticulate package to import flair.models from the Flair NLP Python library.
Ensure that the Python environment is properly set up and the Flair package is installed.
}
\references{
Python equivalent: \preformatted{
from flair.models import *
}
}
