% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_models.R
\name{flair_models.Sequencetagger}
\alias{flair_models.Sequencetagger}
\title{Access Flair's SequenceTagger}
\usage{
flair_models.Sequencetagger()
}
\value{
A Python module (\code{SequenceTagger}) from Flair, which can be utilized to load
and use sequence tagging models.
}
\description{
This function utilizes the reticulate package to import the \code{SequenceTagger}s
from Flair's models in Python, enabling interaction with Flair's sequence
tagging models in an R environment.
}
\details{
The function does not take any parameters and directly returns the
\code{SequenceTagger} when called, which can be used further for sequence tagging
tasks using pre-trained models from Flair.
}
\examples{
\dontrun{
sequence_tagger <- flair_models.sequencetagger()
}

}
\references{
Python equivalent:
\preformatted{
from flair.models import SequenceTagger
}
}
\seealso{
\url{https://github.com/flairNLP/flair} for more information on Flair's
capabilities in NLP and sequence tagging.
}
