% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_models.R
\name{flair_models.TextClassifier}
\alias{flair_models.TextClassifier}
\title{Retrieve TextClassifier from flair.models}
\usage{
flair_models.TextClassifier()
}
\value{
A Python class representing flair.models.TextClassifier.
}
\description{
This function utilizes the reticulate package to directly import TextClassifier from flair.models in the Flair NLP Python library.
Ensure that the Python environment is properly set up and the Flair package is installed.
}
\examples{
# Load the TextClassifier
TextClassifier <- flair_models.TextClassifier()
# Load a pre-trained sentiment model
classifier <- TextClassifier$load('sentiment')

# Create a sentence object
Sentence <- flair_data()$Sentence
sentence <- Sentence("Flair is pretty neat!")

# Predict the sentiment
classifier$predict(sentence)
# Display the sentiment
print(sentence$get_labels())

}
\references{
Python equivalent: \preformatted{
from flair.models import TextClassifier
}
}
