% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_nn.R
\name{flair_nn.Classifier}
\alias{flair_nn.Classifier}
\title{Initializing a Class for Flair Classifier}
\usage{
flair_nn.Classifier()
}
\arguments{
\item{load}{Logical. Indicates if the classifier should be loaded. Default is TRUE.
This parameter is currently defined but not actively used in the function context.}
}
\value{
A Flair Classifier class instance.
}
\description{
This function interfaces with Python via the {reticulate} package
to create a Classifier object from the Flair library.
}
\examples{
\dontrun{
classifier <- flair_nn.Classifier()
}

}
\references{
Python equivalent:
\preformatted{
from flair.nn import Classifier
}
}
