% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapping_python_flair.R
\name{flair_nn.classifier_load}
\alias{flair_nn.classifier_load}
\title{Create a Flair Classifier.load Object}
\usage{
flair_nn.classifier_load(pre_trained_model = "ner")
}
\arguments{
\item{pre_trained_model}{A character string specifying the pre-trained model to use.
This parameter is defined but not used in the current function context.}
}
\value{
A Flair Classifier object.
}
\description{
This function utilizes the `reticulate` package to interface with Python
and create a Classifier object from the Flair library.
}
\examples{
\dontrun{
classifier <- flair_nn_classifier_load("ner")
}
}
\references{
Python equivalent:
\preformatted{
from flair.nn import Classifier
}
}
