% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair.R
\name{flair_rx.default}
\alias{flair_rx.default}
\title{Default S3 method for \code{\link{flair_rx}}.}
\usage{
\method{flair_rx}{default}(x, pattern, code = TRUE, ...)
}
\description{
Default S3 method for \code{\link{flair_rx}}.
}
