% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair.R
\name{flair_rx.with_flair}
\alias{flair_rx.with_flair}
\title{S3 method for \code{\link{with_flair}} objects}
\usage{
\method{flair_rx}{with_flair}(x, pattern, code = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{\link{decorate_code}}.}
}
\value{
An object of class \code{\link{decorate_code}}.
}
\description{
Applies flair to the \code{print_string} attribute.
}
