% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_splitter.R
\name{flair_splitter}
\alias{flair_splitter}
\title{Imort Splitter Module}
\usage{
flair_splitter()
}
\value{
A Python module (`flair.splitter`).
}
\description{
A function to interface with the Python `flair.splitter` module,
specifically utilizing the `SegtokSentenceSplitter` class/method.
}
\examples{
\dontrun{
SegtokSentenceSplitter <- flair_splitter.SegtokSentenceSplitter()
text <- "I am Taiwanese and come from Taiwan"
sentences <- splitter$split(text)
}

}
\references{
Python reference:
\preformatted{
from flair.splitter import SegtokSentenceSplitter
}
}
