% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_splitter.R
\name{flair_splitter.SegtokSentenceSplitter}
\alias{flair_splitter.SegtokSentenceSplitter}
\title{Segtok Sentence Splitter}
\usage{
flair_splitter.SegtokSentenceSplitter()
}
\value{
An instance of the Python class `SegtokSentenceSplitter` from the
`flair.splitter` module.
}
\description{
Interface with the Python `flair.splitter` module to utilize the
`SegtokSentenceSplitter` class/method.
}
\examples{
\dontrun{
splitter <- flair_splitter.SegtokSentenceSplitter()
}

}
\references{
Python equivalent:
\preformatted{
from flair.splitter import SegtokSentenceSplitter
}
}
