% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_lines.R
\name{flair_sublines}
\alias{flair_sublines}
\title{Helper for \code{flair_lines}}
\usage{
flair_sublines(text, nums, lines)
}
\arguments{
\item{text}{Text from knitted source code}

\item{nums}{List of overall lines contained}

\item{lines}{Which lines to highlight}
}
\value{
Text with formatting wrappers
}
\description{
Helper for \code{flair_lines}
}
