% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapping_python_flair.R
\name{flair_trainers}
\alias{flair_trainers}
\title{Import Flair's ModelTrainer in R}
\usage{
flair_trainers()
}
\value{
A Python Module(flair.trainers) object allowing access to Flair's trainers in R.
}
\description{
This function provides R access to Flair's ModelTrainer Python class using the {reticulate} package.
}
\examples{
\dontrun{
trainers <- flair_trainers()
model_trainer <- trainers$ModelTrainer
}
}
\references{
\href{https://github.com/flairNLP/flair}{Flair GitHub}
Python equivalent:
\preformatted{
from flair.trainers import ModelTrainer
}
}
