% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entities.R
\name{get_entities}
\alias{get_entities}
\title{Tagging Named Entities with Flair Standard Models}
\usage{
get_entities(texts, doc_ids, tagger = NULL, ..., language = NULL)
}
\arguments{
\item{texts}{A character vector of texts from which to extract named entities.}

\item{doc_ids}{A character vector of document IDs corresponding to the texts.}

\item{tagger}{An optional tagger object. If NULL, the function will attempt to load the appropriate tagger based on the provided language.}

\item{language}{A character string indicating the language of the texts. Default is "en" (English).}
}
\value{
A data table containing the document ID, entity, and entity label for each named entity extracted from the texts.
}
\description{
Extract named entities from a given text using the flair package.
}
\examples{
\dontrun{
library(reticulate)
library(data.table)

# Load POS tagging model
tagger_post_fast = import("flair.nn")$Classifier$load('pos-fast')

texts <- c("UCD is one of the best university in Ireland.",
           "UCD is good less better than Trinity.",
           "Essex is famous in social science research",
           "Essex is not in Russell Group but it is not bad in politics",
           "Trinity is the oldest one in Ireland.",
           "Trinity is less better than Oxford")
doc_ids <- c("doc1", "doc2", "doc3", "doc4", "doc5", "doc6")

# Load NER model
tagger_ner <- import("flair.nn")$Classifier$load('ner')

results <- get_named_entities_df(texts, doc_ids, tagger_ner)
print(results)}

}
