% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_flair_version}
\alias{get_flair_version}
\title{Retrieve Flair Version}
\usage{
get_flair_version(...)
}
\value{
Character string representing the version of Flair. If Flair is not
installed, this may return \code{NULL} or cause an error.
}
\description{
Gets the version of the installed Flair module in the current
Python environment.
}
\keyword{internal}
