% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ner.R
\name{get_ner}
\alias{get_ner}
\title{Tagging Named Entities with Flair Standard Models}
\usage{
get_ner(texts, doc_ids, tagger = NULL, ..., language = NULL)
}
\arguments{
\item{texts}{A character vector of texts from which to extract named entities.}

\item{doc_ids}{A character vector of document IDs corresponding to the texts.}

\item{tagger}{An optional tagger object. If NULL, the function will attempt to load the appropriate tagger based on the provided language.}

\item{language}{A character string indicating the language of the texts. Default is "en" (English).}
}
\value{
A data table containing the document ID, entity, and entity label for each named entity extracted from the texts.
}
\description{
Extract named entities from a given text using the flair package.
}
\examples{
\dontrun{
texts <- c("UCD is the best university in Ireland.", "Trinity is good too.")
doc_ids <- c("doc1", "doc2")
results <- get_named_entities_df(texts, doc_ids)
print(results)
}

}
