% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate_chunk.R
\name{get_new_globals}
\alias{get_new_globals}
\title{Compares current global chunk options to defaults;
finds options that differ from default.}
\usage{
get_new_globals()
}
\value{
A character vector
}
\description{
Helper for decorate_chunk
}
