% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pos_tags.R
\name{get_pos_tags}
\alias{get_pos_tags}
\title{Tagging part-of-speech tagging with Flair standard model
This function returns a data table of POS tags and other related data for the given texts.}
\usage{
get_pos_tags(texts, doc_ids, tagger = NULL, ..., language = NULL)
}
\arguments{
\item{texts}{A character vector containing texts to be processed.}

\item{doc_ids}{A character vector containing document ids.}

\item{tagger}{A tagger object (default is NULL).}

\item{language}{The language of the texts (default is NULL).}
}
\value{
A data table containing the POS tags and related data.
}
\description{
Tagging part-of-speech tagging with Flair standard model
This function returns a data table of POS tags and other related data for the given texts.
}
\examples{
\dontrun{
# Example usage:
texts <- c("UCD is the best university in Ireland.", "Essex is the worst.")
doc_ids <- c("doc1", "doc2")
result <- get_pos_tags(texts, doc_ids)
print(result)
}
}
