% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sentiment_tag.R
\name{get_sentiment_tag}
\alias{get_sentiment_tag}
\title{Tagging Sentiment with Flair Standard Models
This function takes in texts and their associated document IDs to predict sentiments
using the flair Python library.}
\usage{
get_sentiment_tag(texts, doc_ids, tagger = NULL, ..., language = "en")
}
\arguments{
\item{texts}{A list or vector of texts for which sentiment prediction is to be made.}

\item{doc_ids}{A list or vector of document IDs corresponding to the texts.}

\item{tagger}{An optional flair sentiment model. If NULL (default),
the function loads the default model based on the language.}

\item{language}{A character string indicating the language of the texts.
Currently supports "en" (English), "en-fast" (Fast English), and "de" (German).
Default is "en".}
}
\value{
A data.table with columns: Doc_ID, Sentiment, and Score.
}
\description{
Tagging Sentiment with Flair Standard Models
This function takes in texts and their associated document IDs to predict sentiments
using the flair Python library.
}
