% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sentiments.R
\name{get_sentiments_batch}
\alias{get_sentiments_batch}
\title{Batch Process of Tagging Sentiment with Flair Models}
\usage{
get_sentiments_batch(
  texts,
  doc_ids,
  tagger = NULL,
  ...,
  language = NULL,
  batch_size = 5,
  show.text_id = FALSE,
  gc.active = FALSE
)
}
\arguments{
\item{texts}{A list or vector of texts for which sentiment prediction is
to be made.}

\item{doc_ids}{A list or vector of document IDs corresponding to the texts.}

\item{tagger}{An optional flair sentiment model. If NULL (default),
the function loads the default model based on the language.}

\item{...}{Additional arguments passed to next.}

\item{language}{A character string indicating the language of the texts.
 Currently supports "sentiment" (English), "sentiment-fast" (English), and
"de-offensive-language" (German)}

\item{batch_size}{An integer specifying the number of texts to be processed
at once. It can help optimize performance by leveraging parallel processing.
Default is 5.}

\item{show.text_id}{A logical value. If TRUE, includes the actual text from
which the sentiment was predicted. Default is FALSE.}

\item{gc.active}{A logical value. If TRUE, runs the garbage collector after
processing all texts. This can help in freeing up memory by releasing unused
memory space, especially when processing a large number of texts.
Default is FALSE.}
}
\value{
A \code{data.table} containing three columns:
  \itemize{
    \item \code{doc_id}: The document ID from the input.
    \item \code{sentiment}: Predicted sentiment for the text.
    \item \code{score}: Score for the sentiment prediction.
  }
}
\description{
This function takes in texts and their associated document IDs
to predict sentiments using the flair Python library.
}
\examples{
\dontrun{
library(flaiR)


texts <- c("UCD is one of the best universities in Ireland.",
           "UCD has a good campus but is very far from my apartment in Dublin.",
           "Essex is famous for social science research.",
           "Essex is not in the Russell Group, but it is famous for political science research.",
           "TCD is the oldest university in Ireland.",
           "TCD is similar to Oxford.")

doc_ids <- c("doc1", "doc2", "doc3", "doc4", "doc5", "doc6")

# Load re-trained sentiment ("sentiment") model
tagger_sent <- load_tagger_sentiments('sentiment')

results <- get_sentiments(texts, doc_ids, tagger_sent, batch_size = 3)
print(results)
}

}
