% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gs_score}
\alias{gs_score}
\title{Training Data from : When Do Politicians Grandstand? Measuring Message Politics in Committee Hearings (2021 JOP)}
\format{
A data frame with 3 variables:
\describe{
\item{text}{Hearing text.}
\item{sentimentit_score}{grandstanding score.}
\item{rescaled_gs}{Label indicating whether the text is grandstanding speech:
'1' for grandstanding speech and '0' for non-grandstanding speech. This is a rescaled version
of the sentimentit_score (grandstanding scores) from the original released data.}
}
}
\source{
Data provided by the authors Ju Yeon Park and JOP's Dataverse at
\url{https://dataverse.harvard.edu/file.xhtml?persistentId=doi:10.7910/DVN/GSMBFX/JIHIGH&version=1.0}.
}
\usage{
data("gs_score")
}
\description{
This dataset is from Ju Yeon Park's paper published in the Journal of Politics in 2021,
titled "When Do Politicians Grandstand? Measuring Message Politics in Committee Hearings".
It contains the "Congressional Hearing Dataset: 105th through 114th Congresses",
which is the replication dataset for the paper. The manuscript was accepted for publication
in June 2019. Please cite her paper when using the data.
}
\examples{
\dontrun{
data(gs_score)
head(gs_score)
}
}
