% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_internet}
\alias{has_internet}
\title{Check for Active Internet Connection
This function checks if there's an active internet connection using
the `curl` package. In case of an error or no connection, it will return `FALSE`.}
\usage{
has_internet()
}
\value{
Logical. TRUE if there's an active internet connection, otherwise FALSE.
}
\description{
Check for Active Internet Connection
This function checks if there's an active internet connection using
the `curl` package. In case of an error or no connection, it will return `FALSE`.
}
\keyword{internal}
