% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hatespeech_zh_tw}
\alias{hatespeech_zh_tw}
\title{Training Data from "Political Hate Speech Detection and Lexicon Building: A Study in Taiwan" (IEEE Explore 2022)}
\format{
A data frame with 2 variables:
\describe{
\item{text}{Content of the text.}
\item{label}{Label indicating whether the text is hate speech: '1' for hate speech and '0' for non-hate speech.}
}
}
\source{
Data provided by the authors Chih-Chien Wang, Min-Yuh Day, and Chun-Lian Wu. Available at \url{https://ieeexplore.ieee.org/document/9738642}.
}
\usage{
data("hatespeech_zh_tw")
}
\description{
This dataset is derived sample from the development set of "Political Hate Speech Detection and Lexicon Building: A Study in Taiwan." It contains 1,000 annotated data entries, of which 926 are labeled as '0' (not hate speech) and 74 as '1' (hate speech).

The paper can be accessed at \url{https://ieeexplore.ieee.org/document/9738642}.
}
\examples{
\dontrun{
data(hatespeech_zh_tw)
head(hatespeech_zh_tw)
}
}
