% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{hlt_regexp.default}
\alias{hlt_regexp.default}
\title{Default S3 method for \code{\link{hlt_regexp}}.}
\usage{
\method{hlt_regexp}{default}(.string, pattern, code = TRUE, ...)
}
\description{
Default S3 method for \code{\link{hlt_regexp}}.
}
