% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair.R
\name{import_flair}
\alias{import_flair}
\title{Wrapper for the Flair Python Library}
\usage{
import_flair()
}
\value{
An object that represents the Flair module from Python.
}
\description{
flair: a wrapper to access the Flair library from Python.

Returns:
\itemize{
\item list: A list of method as python module.
}

\describe{
\item{Environment and Configuration:}{
\itemize{
\item \code{os}: Pertains to operating system related functions, such as path
handling, file operations, and more.
\item \code{Path}: From pathlib, used for more convenient file path operations.
\item \code{set_seed}: Functions to set the random seed.
\item \code{hf_set_seed}: Functions to set the random seed.
\item \code{set_proxies}: Used to configure network proxies.
}
}
\item{Data and Data Loading:}{
\itemize{
\item \code{data}: Functions related to data handling and operations.
\item \code{datasets}: Modules or methods to load and handle specific datasets.
\item \code{file_utils}: Utilities for file operations.
}
}
\item{Embeddings and Model Layers:}{
\itemize{
\item \code{embeddings}: About embeddings, including word embeddings,
contextual embeddings, etc.
\item \code{nn}: Related to neural network layers or operations.
\item \code{models}: Different model architectures or structures.
}
}
\item{Training and Optimization:}{
\itemize{
\item \code{trainers}: Related to training models.
\item \code{training_utils}: Utility functions for the training process.
\item \code{optim}: Optimization algorithms, like SGD, Adam.
}
}
\item{Tokenization and Text Processing:}{
\itemize{
\item \code{tokenization}: To break text into tokens.
\item \code{splitter}: For splitting datasets or texts.
}
}
\item{Visualizations and Miscellaneous:}{
\itemize{
\item \code{visual}: Related to visualization.
\item \code{torch}: The main PyTorch library.
\item \code{cache_root}: Related to caching data or models.
}
}
}
}
\details{
This function relies on the \code{reticulate} package to import and
use the Flair module from Python. Ensure you have the Flair Python library
installed in the Python environment being used.
}
\examples{
\dontrun{
flair <- import_flair()
}
}
