% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_python_package}
\alias{install_python_package}
\title{Install a Specific Python Package and Return Its Version}
\usage{
install_python_package(
  package_name,
  package_version = NULL,
  python_path = Sys.which("python3")
)
}
\arguments{
\item{package_name}{The name of the Python package to install.}

\item{package_version}{The version of the Python package to install. If \code{NULL},
the latest version is installed.}

\item{python_path}{The path to the Python interpreter to be used for
installation. If not provided, it defaults to the result of \code{Sys.which("python3")}.}
}
\value{
A list containing the package name, installed version, and the path
to the Python interpreter used for installation.
}
\description{
This function checks for the Python interpreter's location (either
specified by the user or automatically located), compares it with the current
R session's Python setting, installs a specified Python package using the identified
Python interpreter, and returns the package version and installation environment.
}
\examples{
\dontrun{
install_python_package(package_name ="flair", package_version ="0.12")
}
}
