% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorated-print.R
\name{is.decorated}
\alias{is.decorated}
\title{Method check}
\usage{
is.decorated(x)
}
\arguments{
\item{x}{An object}
}
\value{
Whether the object is a \code{decorated} class object.
}
\description{
Method check
}
