% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_flair.R
\name{is.with_flair}
\alias{is.with_flair}
\title{Method check}
\usage{
is.with_flair(x)
}
\arguments{
\item{x}{An object}
}
\value{
Whether the object is a \code{with_flair} class object.
}
\description{
Method check
}
