% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate_code.R
\name{knit_print.decorate_code}
\alias{knit_print.decorate_code}
\title{S3 method for knitting a \code{decorate_code} object}
\usage{
knit_print.decorate_code(x, ...)
}
\description{
S3 method for knitting a \code{decorate_code} object
}
