% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorated-print.R
\name{knit_print.decorated}
\alias{knit_print.decorated}
\title{S3 method for knitting a \code{decorated} object}
\usage{
knit_print.decorated(x, ...)
}
\arguments{
\item{x}{A \code{decorated} object.}

\item{...}{Other \code{knit_print} options}
}
\value{
"as-is" html output, to be rendered when knitted
}
\description{
S3 method for knitting a \code{decorated} object
}
