% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_code.R
\name{knit_print.flair_code}
\alias{knit_print.flair_code}
\title{S3 method for knitting a \code{flair_code} object}
\usage{
knit_print.flair_code(x, ...)
}
\description{
S3 method for knitting a \code{flair_code} object
}
