% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flaired_code.R
\name{knit_print.flaired_code}
\alias{knit_print.flaired_code}
\title{S3 method for knitting a \code{flaired_code} object}
\usage{
knit_print.flaired_code(x, ...)
}
\description{
S3 method for knitting a \code{flaired_code} object
}
