% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_flair.R
\name{knit_print.with_flair}
\alias{knit_print.with_flair}
\title{S3 method for knitting a \code{with_flair} object}
\usage{
knit_print.with_flair(x, ...)
}
\description{
S3 method for knitting a \code{with_flair} object
}
