% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-wrap.R
\name{knitr_wrap}
\alias{knitr_wrap}
\title{weird way to get formatting wrappers
\code{knitr_wrap} returns the output of knitr:::wrap}
\usage{
knitr_wrap(x)
}
\arguments{
\item{x}{evaluated input to wrap}
}
\description{
weird way to get formatting wrappers
\code{knitr_wrap} returns the output of knitr:::wrap
}
