% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate_chunk.R
\name{list_to_strings}
\alias{list_to_strings}
\title{Converts list to vector of strings}
\usage{
list_to_strings(opts_list)
}
\arguments{
\item{opts_list}{A list, presumably of chunk options}
}
\value{
A character vector
}
\description{
Helper for decorate_chunk
}
