% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flaiR_loader.R
\name{load_tagger_pos}
\alias{load_tagger_pos}
\title{Load Flair POS Tagger}
\usage{
load_tagger_pos(language = NULL)
}
\arguments{
\item{language}{A character string indicating the desired language model. If `NULL`,
the function will default to the 'pos-fast' model. Supported language models include:
\itemize{
  \item "pos" - General POS tagging
  \item "pos-fast" - Faster POS tagging
  \item "upos" - Universal POS tagging
  \item "upos-fast" - Faster Universal POS tagging
  \item "pos-multi" - Multi-language POS tagging
  \item "pos-multi-fast" - Faster Multi-language POS tagging
  \item "ar-pos" - Arabic POS tagging
  \item "de-pos" - German POS tagging
  \item "de-pos-tweets" - German POS tagging for tweets
  \item "da-pos" - Danish POS tagging
  \item "ml-pos" - Malayalam POS tagging
  \item "ml-upos" - Malayalam Universal POS tagging
  \item "pt-pos-clinical" - Clinical Portuguese POS tagging
  \item "pos-ukrainian" - Ukrainian POS tagging
}}
}
\value{
A Flair POS tagger model corresponding to the specified (or default) language.
}
\description{
This function loads the POS (part-of-speech) tagger model for a specified language
using the Flair library. If no language is specified, it defaults to 'pos-fast'.
}
\examples{
\dontrun{
tagger <- load_tagger_pos("pos-fast")
}
}
