% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_tagger_sentiments}
\alias{load_tagger_sentiments}
\title{Load a Sentiment or Language Tagger Model from Flair}
\usage{
load_tagger_sentiments(language)
}
\arguments{
\item{language}{Character string specifying the language model to load.
Can be one of "sentiment", "sentiment-fast", or "de-offensive-language".
Defaults to "sentiment" if not provided.}
}
\value{
An object of the loaded Flair model.
}
\description{
This function loads a pre-trained sentiment or language tagger
from the Flair library.  It allows you to specify the model language you wish to load.
}
\examples{
\dontrun{
  tagger <- load_tagger_sentiments("sentiment")
}

}
