% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate_code.R
\name{print.decorate_code}
\alias{print.decorate_code}
\title{S3 method for printing a \code{decorate_code} object}
\usage{
\method{print}{decorate_code}(x, ...)
}
\description{
Prints nothing directly; \code{decorate_code} objects should be seen and not heard.
}
\details{
However, when \code{eval = TRUE} was specified, the original source code should be run and output printed.
}
