% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_flair.R
\name{print.decorated}
\alias{print.decorated}
\title{When run interactively, a \code{decorated} object should preview the
flaired source code in the viewer pane. (Only if in RStudio.)}
\usage{
\method{print}{decorated}(x, ...)
}
\arguments{
\item{x}{A \code{decorated} object.}

\item{...}{Other \code{print} options}
}
\value{
None
}
\description{
When run interactively, a \code{decorated} object should preview the
flaired source code in the viewer pane. (Only if in RStudio.)
}
