% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_code.R
\name{print.flair_code}
\alias{print.flair_code}
\title{S3 method for printing a \code{flair_code}}
\usage{
\method{print}{flair_code}(x, ...)
}
\description{
Prints nothing; \code{flair_code} objects should be seen and not heard.
}
\details{
If the \code{flair_code} object was created by inputting a string, we should run that code and print any output.
}
