% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flaired_code.R
\name{print.flaired_code}
\alias{print.flaired_code}
\title{S3 method for printing a \code{flaired_code}}
\usage{
\method{print}{flaired_code}(x, ...)
}
\description{
Prints nothing; \code{flaired_code} objects should be seen and not heard.
}
\details{
If the \code{flaired_code} object was created by inputting a string, we should run that code and print any output.
}
