% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-scope.R
\name{scope_and_run}
\alias{scope_and_run}
\title{Runs code from string, in parent environment}
\usage{
scope_and_run(.code_string)
}
\arguments{
\item{.code_string}{A string containing runnable R code.}
}
\value{
Nothing; side effects in environment only.
}
\description{
Shortcut function to rescope a code string and then run (but not print).
Looks for object assignments of the form \code{foo <-} and rescopes to
\code{foo <<-}, then evaluates code string.
}
