% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_flair_cache}
\alias{show_flair_cache}
\title{Show Flair Cache Preloaed flair's Directory}
\usage{
show_flair_cache()
}
\value{
A data frame containing the file paths of the contents in the flair
cache directory. If the directory does not exist or is empty, NULL is returned.
}
\description{
This function lists the contents of the flair cache directory
and returns them as a data frame.
}
\examples{
\dontrun{
show_flair_cache()
}
}
