% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate_chunk.R
\name{src_to_list}
\alias{src_to_list}
\title{Takes plain text of knitted code and converts it to a list, in which code
sources have the class \code{source}.}
\usage{
src_to_list(knitted)
}
\arguments{
\item{knitted}{Text of knitted code}
}
\value{
A list with code sources and knitted output.
}
\description{
Takes plain text of knitted code and converts it to a list, in which code
sources have the class \code{source}.
}
