% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{statements}
\alias{statements}
\title{Sampled Grandstanding Text}
\format{
A data frame with 3 variables:
\describe{
\item{Type}{Grandstanding types}
\item{Statement}{Grandstanding texts}
}
}
\usage{
data("statements")
}
\description{
Sampled Grandstanding Text
}
\examples{
\dontrun{
data(statements)
head(statements)
}
}
