% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uninstall_python_package}
\alias{uninstall_python_package}
\title{Uninstall a Python Package}
\usage{
uninstall_python_package(package_name, python_path = Sys.which("python3"))
}
\arguments{
\item{package_name}{The name of the Python package to uninstall.}

\item{python_path}{The path to the Python executable. If not provided, it uses the system's default Python path.}
}
\value{
Invisibly returns TRUE if the package is successfully uninstalled, otherwise it stops with an error message.
}
\description{
`uninstall_python_package` function uninstalls a specified Python
package using the system's Python installation. It checks if Python is
installed and accessible, then proceeds to uninstall the package. Finally,
`uninstall_python_package` verifies that the package has been successfully uninstalled.
}
\examples{
\dontrun{
uninstall_python_package("numpy")
}
}
