% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{upgrade_python_packages}
\alias{upgrade_python_packages}
\title{Upgrade specific Python packages}
\usage{
upgrade_python_packages()
}
\value{
A numeric vector indicating the success or failure for each command.
        A return value of 0 typically indicates success.
}
\description{
This function upgrades pip, transformers, flair, and numpy using
pip.
}
\keyword{internal}
