% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_flair.R
\name{with_flair}
\alias{with_flair}
\title{Create a \code{with_flair} object}
\usage{
with_flair(x)
}
\arguments{
\item{x}{A text object or code chunk label}
}
\value{
An object of class \code{with_flair}
}
\description{
(The preferred function is \code{\link{decorate}})
}
