% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-wrap.R
\name{wrap_html}
\alias{wrap_html}
\title{Wraps text in html for formatting}
\usage{
wrap_html(x, opts_list)
}
\arguments{
\item{x}{The string to be wrapped}

\item{opts_list}{The html options to include, in named vector form}

\item{type}{The style of display, defaults to "html"  (currently nothing else is supported, sorry)}
}
\description{
\code{wrap_html} returns a string with styling wrappers from a list of style tag names and values.
}
