% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_and_gc}
\alias{check_and_gc}
\title{Perform Garbage Collection Based on Condition}
\usage{
check_and_gc(gc.active)
}
\arguments{
\item{gc.active}{A logical value indicating whether or not to
activate garbage collection.}
}
\value{
A message indicating that garbage collection was performed
if \code{gc.active} was \code{TRUE}. Otherwise, no action is taken or message is
displayed.
}
\description{
This function checks the value of \code{gc.active} to determine whether
or not to perform garbage collection. If \code{gc.active} is \code{TRUE},
the function will perform garbage collection and then send a
message indicating the completion of this process.
}
\keyword{internal}
