% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_device}
\alias{check_device}
\title{Check the Device for cccelerating PyTorch}
\usage{
check_device(device)
}
\arguments{
\item{device}{Character. The device to be set for PyTorch.}
}
\description{
This function verifies if the specified device is available for PyTorch.
If CUDA is not available, a message is shown. Additionally, if the system
is running on a Mac M1, MPS will be used instead of CUDA.
}
\keyword{internal}
