% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_device}
\alias{check_device}
\title{Check the specified device for PyTorch}
\usage{
check_device(device)
}
\arguments{
\item{device}{Character. The device to be set for PyTorch.}
}
\description{
This function verifies if the specified device is available for PyTorch.
If CUDA is not available, a message is shown.
}
\keyword{internal}
