% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_flair_installed}
\alias{check_flair_installed}
\title{Check if flair is installed in the Python environment.}
\usage{
check_flair_installed()
}
\value{
Logical. TRUE if flair is installed, otherwise it will stop with an error message.
}
\description{
This function checks if the `flair` Python module is available in the
current R session's Python environment. If `flair` is not installed,
or if Python is not available, an error will be raised.
}
