% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_python_installed}
\alias{check_python_installed}
\title{Check for Available Python Installation}
\usage{
check_python_installed()
}
\value{
Logical. `TRUE` if Python is installed, `FALSE` otherwise. Additionally, if installed, the path to the Python installation is printed.
}
\description{
This function checks if any environment is installed on the R system.
}
