% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_flair_env}
\alias{create_flair_env}
\title{Create and setup the `r-reticulate` environment for Flair using reticulate.}
\usage{
create_flair_env(...)
}
\value{
NULL. The function is used for its side effects of setting up the conda environment.
}
\description{
This function ensures that the `flair` Python module is installed in the
`r-reticulate` conda environment created by the `reticulate` R package.
If the `flair` module is already installed, the function prints its version
and exits without making any changes.
}
