% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_flair_env}
\alias{create_flair_env}
\title{Create or use Python environment for Flair}
\usage{
create_flair_env(env = "r-reticulate")
}
\arguments{
\item{env}{The name of the conda environment to be used or created (default is "r-reticulate").}
}
\value{
Nothing is returned. The function primarily ensures that the Python library Flair is installed and available.
}
\description{
This function checks whether the Flair Python library is installed in the current Python environment.
If it is not, it attempts to install it either in the current conda environment or creates a new one.
}
\examples{
\dontrun{
  create_flair_env()
}
}
