% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate}
\alias{decorate}
\title{Builds a \code{\link{with_flair}} object from either a code chunk or a
string object containing source code.}
\usage{
decorate(x, ...)
}
\arguments{
\item{x}{A string, containing either a chunk label or R code.}

\item{...}{Chunk options to pass along}
}
\value{
A \code{with_flair} object.
}
\description{
\code{decorate} does its best to guess if it has been given a code string or
a chunk name, based on the presence of special characters.
}
\seealso{
\code{\link{decorate_chunk}}, \code{\link{decorate_code}}
}
