% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{#' @export .onLoad
.onLoad <- function(...) {
  python_version <- reticulate::py_config()$version
  if (python_version < "3.7") {
    stop("flaiR requires Python 3.7 or higher.")
  }
}}
\usage{
.onAttach(...)
}
\description{
#' @export .onLoad
.onLoad <- function(...) {
  python_version <- reticulate::py_config()$version
  if (python_version < "3.7") {
    stop("flaiR requires Python 3.7 or higher.")
  }
}
}
