% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{.onAttach Function for the flaiR Package}
\usage{
.onAttach(...)
}
\description{
This function is called when the flaiR package is loaded. \
It provides messages detailing the versions of Python and Flair being used, a
s well as other package details.
}
\keyword{internal}
