% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{flaiR-package}
\alias{flaiR}
\alias{flaiR-package}
\title{flaiR: An R Wrapper for Accessing Flair NLP}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides an R interface to the Flair Natural Language Processing library in Python. This package allows R users to access Flair's NLP important features including named entity recognition, part-of-speech tagging, text embeddings and word embeddings. The package handles the Python dependencies management and provides R-friendly wrappers for Flair's functionality. Users need Python (>= 3.9.0) with PyTorch (>= 2.2.0), Transformers (>= 4.37.2), and Flair (>= 0.11.3) installed in their system. The package automatically checks and validates the Python environment during loading, ensuring all required dependencies are properly configured.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://davidycliao.github.io/flaiR}
  \item Report bugs at \url{https://github.com/davidycliao/flaiR/issues}
}

}
\author{
\strong{Maintainer}: Yen-Chieh Liao \email{davidycliao@gmail.com} [contributor]

Authors:
\itemize{
  \item Akbik Alan \email{alan.akbik@zalando.de} [contributor]
  \item Blythe Duncan \email{blythe.duncan@zalando.de} [contributor]
  \item Vollgraf Roland \email{vollgraf.roland@zalando.de} [contributor]
}

}
\keyword{internal}
