% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{flaiR-package}
\alias{flaiR}
\alias{flaiR-package}
\title{flaiR: An R Wrapper for Accessing FLAIR}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a suite of intuitive and flexible wrapper tools tailored for accessing FLAIR in Python data analysis. The package is an R wrapper designed to provide access to the primary features of Flair NLP.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://davidycliao.github.io/flaiR}
  \item Report bugs at \url{https://github.com/davidycliao/flaiR/issues}
}

}
\author{
\strong{Maintainer}: Yen-Chieh Liao \email{davidyclaio@gmail.com} [contributor]

Authors:
\itemize{
  \item Stefan Müller \email{stefan.mueller@ucd.ie} [contributor]
  \item Akbik Alan \email{alan.akbik@zalando.de} [contributor]
  \item Blythe Duncan \email{blythe.duncan@zalando.de} [contributor]
  \item Vollgraf Roland \email{vollgraf.roland@zalando.de} [contributor]
}

Other contributors:
\itemize{
  \item Sohini Timbadia [contributor]
}

}
\keyword{internal}
