% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair-package.R
\docType{package}
\name{flair-package}
\alias{flair-package}
\title{A package for adding flair to your displayed source code.}
\description{
This package provides ...
}
\section{Flair functions}{


The \code{flair_*} functions add decorative formatting to source code.
}

\section{Decorate functions}{


The \code{decorate} functions execute source code and simultaneously prepare
it to have flair added, to display when knitted.
}

