% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair.R
\name{flair}
\alias{flair}
\title{Full Flair Module}
\usage{
flair()
}
\value{
Flair Python module
}
\description{
This function is a wrapper for the Flair Python library.
}
\examples{
\dontrun{
flair <- flair()
}
}
\references{
Python equivalent: \preformatted{
import flair
}
}
