% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{flaiR}
\alias{flaiR}
\alias{flaiR-package}
\title{FlaiR: R Interface to Flair NLP}
\description{
FlaiR provides an R interface to the Flair NLP library, enabling natural
language processing tasks in R using Python's Flair library.
}
