% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flar_data.R
\name{flair_data}
\alias{flair_data}
\title{Flair Data Import Function}
\usage{
flair_data()
}
\value{
A reference to the `data` module of the Flair Python package.
}
\description{
This function imports and returns the `data` module from the
Flair package in Python.
}
\references{
From Flair's Python library: \code{from flair.data import *}
}
