% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flar_data.R
\name{flair_data}
\alias{flair_data}
\title{Import flair.data Module}
\usage{
flair_data(load = TRUE)
}
\value{
A Python module (`flair.splitter`).
}
\description{
A function to interface with the Python `flair.splitter` module,
specifically to load the `SegtokSentenceSplitter` class/method.
}
\examples{
\dontrun{
Sentence <- flair_data()$Sentence
sentences <- Sentence("I am Taiwanese and come from Taiwan.")
}

}
\references{
Python reference:
\preformatted{
from flair.data import Sentence
}
}
