% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapping_python_flair.R
\name{flair_data.sentence}
\alias{flair_data.sentence}
\title{Create a Flair Sentence Object}
\usage{
flair_data.sentence(sentence_text)
}
\arguments{
\item{sentence_text}{A character string to be converted into a Flair Sentence object.}
}
\value{
A Flair Sentence object.
}
\description{
This function uses the `reticulate` package to interface with Python and
create a Flair Sentence object.
}
\examples{
\dontrun{
flair_data_sentence("The quick brown fox jumps over the lazy dog.")
}
}
\references{
Python equivalent:
\preformatted{
from flair.data import Sentence
sentence = Sentence("The quick brown fox jumps over the lazy dog.")
}
}
