% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flar_data.R
\name{flair_data.Sentence}
\alias{flair_data.Sentence}
\title{Create a Flair Sentence Object}
\usage{
flair_data.Sentence(sentence_text)
}
\arguments{
\item{sentence_text}{A character string to be converted into a Flair Sentence object.}
}
\value{
A Flair Sentence object.
}
\description{
This function uses the {reticulate} package to interface with Python and
create a Flair Sentence object.
}
\examples{
\dontrun{
flair_data.sentence("The quick brown fox jumps over the lazy dog.")
}

}
\references{
Python equivalent: \preformatted{
from flair.data import Sentence
sentence = Sentence("The quick brown fox jumps over the lazy dog.")
}
}
