% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_embeddings.R
\name{flair_embeddings}
\alias{flair_embeddings}
\title{Initialization of of Flair Embeddings Modules}
\usage{
flair_embeddings()
}
\value{
The \code{flair.embeddings} module from Flair.
}
\description{
This function imports and returns the \code{flair.embeddings} module from Flair.
It provides a convenient R interface to the Flair library's embedding functionalities.
}
\examples{
\dontrun{
flair_embeddings <- flair_embeddings()$FlairEmbeddings
OpenAIGPTEmbeddings <- flair_embeddings()$OpenAIGPTEmbeddings
StackedEmbeddings <- flair_embeddings()$StackedEmbeddings
TransformerDocumentEmbeddings <- flair_embeddings()$TransformerDocumentEmbeddings
TransformerWordEmbeddings <- flair_embeddings()$TransformerWordEmbeddings
RoBERTaEmbeddings <- flair_embeddings()$RoBERTaEmbeddings
TransformerOnnxDocumentEmbeddings <- flair_embeddings()$TransformerOnnxDocumentEmbeddings
SentenceTransformerDocumentEmbeddings <- flair_embeddings()$SentenceTransformerDocumentEmbeddings
PooledFlairEmbeddings <- flair_embeddings()$PooledFlairEmbeddings
}

}
\references{
In Python's Flair library:
\code{
from flair.embeddings import FlairEmbeddings
}
}
