% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapping_python_flair.R
\name{flair_embeddings.TransformerDocumentEmbeddings}
\alias{flair_embeddings.TransformerDocumentEmbeddings}
\title{TransformerDocumentEmbeddings Function}
\usage{
flair_embeddings.TransformerDocumentEmbeddings(
  pre_trained = "bert-base-uncased"
)
}
\arguments{
\item{pre_trained}{A string specifying the name of a pre-trained transformer model.}
}
\value{
An instance of the TransformerDocumentEmbeddings model from the Flair library.
}
\description{
This function initializes and returns a Transformer Document Embedding model from the Flair library.
It takes a pre-trained model name as an argument and returns the respective embedding model.
}
\examples{
\dontrun{
sentence <- flair_data.sentence('Your sentence here.')
embedding <- flair_embeddings.TransformerDocumentEmbeddings(pre_trained = "bert-base-uncased")
embedding.embed(sentence)
print(sentence) }

}
\references{
In Python's Flair library:
\code{
from flair.embeddings import TransformerDocumentEmbeddings
from flair.data import Sentence
embedding = TransformerDocumentEmbeddings('bert-base-uncased')
sentence = Sentence('flaiR is R wrapper')
embedding.embed(sentence)
}
}
