% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapping_python_flair.R
\name{flair_embeddings.WordEmbeddings}
\alias{flair_embeddings.WordEmbeddings}
\title{Create a Flair WordEmbeddings Object}
\usage{
flair_embeddings.WordEmbeddings(pre_trained = "glove")
}
\arguments{
\item{pre_trained}{A character string specifying the pre-trained model to use.
Defaults to "`glove`".}
}
\value{
A Flair WordEmbeddings object.
}
\description{
This function interfaces with Python via {reticulate} to create
a `WordEmbeddings` object using the Flair library.
}
\examples{
\dontrun{
sentence = Sentence('flaiR is R wrapper')
embedding <- word_embeddings("glove")
}

}
\references{
Python equivalent:
\preformatted{
from flair.data import Sentence
from flair.embeddings import WordEmbeddings
embedding = WordEmbeddings('glove')
sentence = Sentence('flaiR is R wrapper')
embedding.embed(sentence)
print(sentence)
}
}
