% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapping_python_flair.R
\name{flair_nn}
\alias{flair_nn}
\title{Import Flair's Neural Network Module}
\usage{
flair_nn(pre_trained)
}
\arguments{
\item{pre_trained}{A parameter (currently not being used in the function body).}
}
\value{
A reference to Flair's neural network module.
}
\description{
This function imports the neural network module from the
Flair library.
}
\examples{
\dontrun{
  flair_nn <- flair_nn(pre_trained = TRUE)
  Classifier <- flair_nn$Classifier
}
}
