% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_nn.R
\name{flair_nn}
\alias{flair_nn}
\title{Import Flair's Neural Network Module}
\usage{
flair_nn()
}
\value{
A reference to Flair's neural network module.
}
\description{
This function imports the neural network module from the
Flair library.
}
\examples{
\dontrun{
  flair_nn <- flair_nn(load = TRUE)
  Classifier <- flair_nn$Classifier
}
}
