% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_nn.R
\name{flair_nn}
\alias{flair_nn}
\title{Import Flair's Neural Network Module}
\usage{
flair_nn()
}
\value{
A reference to Flair's neural network module (\code{flair.nn}).
}
\description{
This function provides an interface to the \code{flair.nn} module from the Flair library.

The \code{flair.nn} module encompasses various sub-modules such as:
\itemize{
\item decoder
\item distance
\item dropout
\item loss
\item model
\item multitask
\item recurrent
\item Model
\item Classifier
\item PrototypicalDecoder
\item LockedDropout
\item WordDropout
}
}
\examples{
\dontrun{
  flair_nn_module <- flair_nn()
  Classifier <- flair_nn_module$Classifier
}
}
