% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair.R
\name{flair_rx.default}
\alias{flair_rx.default}
\title{Default S3 method for \code{\link{flair_rx}}.}
\usage{
\method{flair_rx}{default}(x, pattern, before = NULL, after = NULL, ...)
}
\arguments{
\item{x}{A string or \code{\link{with_flair}} object}

\item{pattern}{A pattern to match.  By default, this is a fixed pattern;
use \code{flair_rx} for regular expressions.}

\item{before}{String giving specific html tags to insert before matched text.}

\item{after}{String giving specific html tags to insert after matched text.}

\item{...}{Formatting style options, passed to \code{\link{txt_style}}}
}
\description{
Default S3 method for \code{\link{flair_rx}}.
}
