% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair.R
\name{flair_rx.with_flair}
\alias{flair_rx.with_flair}
\title{S3 method for \code{\link{with_flair}} objects}
\usage{
\method{flair_rx}{with_flair}(x, pattern, before = NULL, after = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{\link{with_flair}}.}
}
\value{
An object of class \code{\link{with_flair}}.
}
\description{
Applies flair to the source code.
}
