% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_splitter.R
\name{flair_splitter}
\alias{flair_splitter}
\title{Import flair.splitter Module in R}
\usage{
flair_splitter()
}
\value{
A Python module (`flair.splitter`).
}
\description{
A function to interface with the Python \code{flair.splitter} module. This function provides access to various
sentence splitting strategies implemented in the Flair library:
\itemize{
  \item \code{NoSentenceSplitter}: Treats the entire text as a single sentence without splitting it.
  \item \code{SegtokSentenceSplitter}: Uses the \code{segtok} library to split text into sentences.
  \item \code{SpacySentenceSplitter}: Uses the \code{spaCy} library for sentence splitting.
  \item \code{TagSentenceSplitter}: Assumes specific tags in the text to indicate sentence boundaries.
}
}
\examples{
\dontrun{
SegtokSentenceSplitter <- flair_splitter$SegtokSentenceSplitter()
text <- "I am Taiwanese and come from Taiwan"
sentences <- splitter$split(text)
}

}
\references{
Python reference for `SegtokSentenceSplitter`:
\preformatted{
from flair.splitter import *
}
Additional references for the other classes can be found within the Flair library documentation.
\href{https://github.com/flairNLP/flair/blob/master/flair/splitter.py}{Flair GitHub}
}
