% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_splitter.R, R/flar_data.R
\name{flair_splitter}
\alias{flair_splitter}
\title{Imort Splitter Module}
\usage{
flair_splitter(load = TRUE)

flair_splitter(load = TRUE)
}
\value{
A Python module (`flair.splitter`).

A Python module (`flair.splitter`).
}
\description{
A function to interface with the Python `flair.splitter` module,
specifically utilizing the `SegtokSentenceSplitter` class/method.

A function to interface with the Python `flair.splitter` module,
specifically to load the `SegtokSentenceSplitter` class/method.
}
\examples{
\dontrun{
SegtokSentenceSplitter <- flair_splitter.SegtokSentenceSplitter()
text <- "I am Taiwanese and come from Taiwan"
sentences <- splitter$split(text)
}

\dontrun{
splitter_module <- flair_splitter()
SegtokSentenceSplitter <- splitter_module$SegtokSentenceSplitter
text <- "I am Taiwanese and come from Taiwan."
splitter_instance <- SegtokSentenceSplitter()
sentences <- splitter_instance$split(text)
}

}
\references{
Python reference:
\preformatted{
from flair.splitter import SegtokSentenceSplitter
}

Python reference:
\preformatted{
from flair.splitter import SegtokSentenceSplitter
}
}
