% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flair_splitter.R
\name{flair_splitter.SegtokSentenceSplitter}
\alias{flair_splitter.SegtokSentenceSplitter}
\title{Segtok Sentence Splitter}
\usage{
flair_splitter.SegtokSentenceSplitter()
}
\value{
A Python module (`flair.splitter`).
}
\description{
A function to interface with the Python `flair.splitter` module,
specifically utilizing the `SegtokSentenceSplitter` class/method.
}
\examples{
\dontrun{
splitter <- flair_splitter.SegtokSentenceSplitter()
}

}
\references{
Python reference:
\preformatted{
from flair.splitter import SegtokSentenceSplitter
}
}
