% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entities.R
\name{get_entities}
\alias{get_entities}
\title{Tagging Named Entities with Flair Standard Models}
\usage{
get_entities(texts, doc_ids, tagger = NULL, ..., language = NULL)
}
\arguments{
\item{texts}{A character vector of texts from which to extract named entities.}

\item{doc_ids}{A character vector of document IDs corresponding to the texts.}

\item{tagger}{An optional tagger object. If NULL, the function will attempt to load the appropriate tagger based on the provided language.}

\item{...}{Additional arguments passed to next.}

\item{language}{A character string indicating the language of the texts. Default is "en" (English).}
}
\value{
A data table containing the document ID, entity, and entity label for each named entity extracted from the texts.
}
\description{
Extract named entities from a given text using the flair package.
}
\examples{
\dontrun{
library(reticulate)
library(fliaR)

texts <- c("UCD is one of the best universities in Ireland.",
           "UCD has a good campus but is very far from my apartment in Dublin.",
           "Essex is famous for social science research.",
           "Essex is not in the Russell Group, but it is famous for political science research.",
           "TCD is the oldest university in Ireland.",
           "TCD is similar to Oxford.")

# Load NER ("ner") model
tagger_ner <- import("flair.nn")$Classifier$load('ner')

results <- get_entities(texts, doc_ids, tagger_ner)
print(results)}

}
