% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pos.R
\name{get_pos}
\alias{get_pos}
\title{Tagging Part-of-Speech Tagging with Flair Standard Models
This function returns a data table of POS tags and other related data for the given texts.}
\usage{
get_pos(texts, doc_ids, tagger = NULL, ..., language = NULL)
}
\arguments{
\item{texts}{A character vector containing texts to be processed.}

\item{doc_ids}{A character vector containing document ids.}

\item{tagger}{A tagger object (default is NULL).}

\item{language}{The language of the texts (default is NULL).}
}
\value{
A `data.table` containing the following columns:
\describe{
  \item{\code{doc_id}}{The document identifier corresponding to each text.}
  \item{\code{token_id}}{The token number in the original text, indicating the position of the token.}
  \item{\code{text_id}}{The actual text input passed to the function.}
  \item{\code{token}}{The individual word or token from the text that was POS tagged.}
  \item{\code{tag}}{The part-of-speech tag assigned to the token by the Flair library.}
  \item{\code{precision}}{A confidence score (numeric) for the assigned POS tag.}
}
}
\description{
Tagging Part-of-Speech Tagging with Flair Standard Models
This function returns a data table of POS tags and other related data for the given texts.
}
\examples{
\dontrun{
library(reticulate)
library(data.table)
# Load Part-of-Speech  ("pos-fast") model
tagger_pos_fast = import("flair.nn")$Classifier$load('pos-fast')
texts <- c("UCD is one of the best universities in Ireland.",
           "UCD has a good campus but is very far from my apartment in Dublin.",
           "Essex is famous for social science research.",
           "Essex is not in the Russell Group, but it is famous for political science research.",
           "TCD is the oldest university in Ireland.",
           "TCD is similar to Oxford.")
doc_ids <- c("doc1", "doc2", "doc3", "doc4", "doc5", "doc6")

get_pos(texts, doc_ids, tagger_pos_fast)
}
}
