% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sentiments.R
\name{get_sentiments}
\alias{get_sentiments}
\title{Tagging Sentiment with Flair Standard Models
This function takes in texts and their associated document IDs to predict sentiments
using the flair Python library.}
\usage{
get_sentiments(texts, doc_ids, tagger = NULL, ..., language = NULL)
}
\arguments{
\item{texts}{A list or vector of texts for which sentiment prediction is to be made.}

\item{doc_ids}{A list or vector of document IDs corresponding to the texts.}

\item{tagger}{An optional flair sentiment model. If NULL (default),
the function loads the default model based on the language.}

\item{language}{A character string indicating the language of the texts.
Currently supports "en" (English), "en-fast" (Fast English), and "de" (German).
Default is "en".}
}
\value{
A \code{data.table} containing three columns:
  \itemize{
    \item \code{doc_iid}: The document ID from the input.
    \item \code{sentiment}: Predicted sentiment for the text.
    \item \code{score}: score for the sentiment prediction.
  }
}
\description{
Tagging Sentiment with Flair Standard Models
This function takes in texts and their associated document IDs to predict sentiments
using the flair Python library.
}
\examples{
\dontrun{
library(reticulate)
library(data.table)


texts <- c("UCD is one of the best universities in Ireland.",
           "UCD is good but not quite as good as Trinity.",
           "Essex is famous for social science research.",
           "Essex is not in the Russell Group, but it is famous for political science research.",
           "TCD is the oldest university in Ireland.",
           "TCD is similar to Oxford.")

doc_ids <- c("doc1", "doc2", "doc3", "doc4", "doc5", "doc6")

# Load re-trained sentiment ("sentiment") model
tagger_sent <- import("flair.nn")$Classifier$load('sentiment')

results <- get_sentiments(texts, doc_ids, tagger_sent)
print(results)
}

}
