% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{highlight_text}
\alias{highlight_text}
\title{Highlight Entities with Specified Colors and Tag}
\usage{
highlight_text(text, entities_mapping, font_family = "Arial")
}
\arguments{
\item{text}{A character string containing the text to highlight.}

\item{entities_mapping}{A named list of lists, with each sub-list containing:
\itemize{
  \item \code{words}: A character vector of words to highlight.
  \item \code{background_color}: A character string specifying the CSS color for the highlight background.
  \item \code{font_color}: A character string specifying the CSS color for the highlighted text.
  \item \code{label}: A character string specifying a label to append after each highlighted word.
  \item \code{label_color}: A character string specifying the CSS color for the label text.
}}

\item{font_family}{A character string specifying the CSS font family for
the highlighted text and label. Default is "Arial".}
}
\value{
An HTML object containing the text with highlighted entities.
}
\description{
This function highlights specified entities in a text string
with specified background colors, font colors, and optional labels.
Additionally, it allows setting a specific font type for highlighted text.
}
\examples{
\dontrun{
  entities_mapping <- list(
    ORG = list(words = c("ORG1", "ORG2"),
               background_color = "pink", font_color = "black",
               label = "ORG", label_color = "pink")
  )
  highlighted_text <- highlight_text("Example text with ORG1 and ORG2.", entities_mapping)
}

}
