% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_tagger_ner}
\alias{load_tagger_ner}
\title{Load the Named Entity Recognition (NER) Tagger}
\usage{
load_tagger_ner(language)
}
\arguments{
\item{language}{Character string indicating the desired language for the NER tagger.
Supported languages include "en", "de", "fr", "nl", "da", and "ar".}
}
\value{
An instance of the Flair SequenceTagger for the specified language.
}
\description{
A helper function to load the appropriate tagger based on the provided language.
This function supports a variety of languages/models.
}
\examples{
# Load the English NER tagger
tagger_en <- load_tagger_ner("en")

}
