% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_tagger_ner}
\alias{load_tagger_ner}
\title{Load the Named Entity Recognition (NER) Tagger}
\usage{
load_tagger_ner(language = NULL)
}
\arguments{
\item{language}{Character string indicating the desired language for the NER tagger.
Supported languages include "en", "de", "fr", "nl", "da", and "ar".}
}
\value{
An instance of the Flair SequenceTagger for the specified language.
}
\description{
A helper function to load the appropriate tagger based on the provided language.
This function supports a variety of languages/models.
}
\details{
Supported languages and their corresponding codes are:
* "en" - English: `ner`
* "de" - German: `de-ner`
* "fr" - French: `fr-ner`
* "nl" - Dutch: `nl-ner`
* "da" - Danish: `da-ner`
* "ar" - Arabic: `ar-ner`
}
\examples{
# Load the English NER tagger
tagger_en <- load_tagger_ner("en")

}
