% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flaiR_loader.R
\name{load_tagger_ner}
\alias{load_tagger_ner}
\title{Load the Named Entity Recognition (NER) Tagger}
\usage{
load_tagger_ner(language = NULL)
}
\arguments{
\item{language}{A character string indicating the desired language for the NER tagger.
If \code{NULL}, the function will default to the 'pos-fast' model.
Supported languages and their models include:
\itemize{
\item \code{"en"} - English NER tagging (\code{ner})
\item \code{"de"} - German NER tagging (\code{de-ner})
\item \code{"fr"} - French NER tagging (\code{fr-ner})
\item \code{"nl"} - Dutch NER tagging (\code{nl-ner})
\item \code{"da"} - Danish NER tagging (\code{da-ner})
\item \code{"ar"} - Arabic NER tagging (\code{ar-ner})
\item \code{"ner-fast"} - English NER fast model (\code{ner-fast})
\item \code{"ner-large"} - English NER large mode (\code{ner-large})
\item \code{"de-ner-legal"} - NER (legal text) (\code{de-ner-legal})
\item \code{"nl"} - Dutch NER tagging (\code{nl-ner})
\item \code{"da"} - Danish NER tagging (\code{da-ner})
\item \code{"ar"} - Arabic NER tagging (\code{ar-ner})
}}
}
\value{
An instance of the Flair SequenceTagger for the specified language.
}
\description{
A helper function to load the appropriate tagger based on the provided language.
This function supports a variety of languages/models.
}
\examples{
# Load the English NER tagger
tagger_en <- load_tagger_ner("en")

}
