% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_tagger_ner}
\alias{load_tagger_ner}
\title{Load the Named Entity Recognition (NER) Tagger}
\usage{
load_tagger_ner(language = NULL)
}
\arguments{
\item{language}{A character string indicating the desired language for the NER tagger.
If `NULL`, the function will default to the 'pos-fast' model.
Supported languages and their models include:
\itemize{
  \item "en" - English NER tagging (`ner`)
  \item "de" - German NER tagging (`de-ner``)
  \item "fr" - French NER tagging (`fr-ner`)
  \item "nl" - Dutch NER tagging (`nl-ner`)
  \item "da" - Danish NER tagging (`da-ner`)
  \item "ar" - Arabic NER tagging (`ar-ner`)
}}
}
\value{
An instance of the Flair SequenceTagger for the specified language.
}
\description{
A helper function to load the appropriate tagger based on the provided language.
This function supports a variety of languages/models.
}
\examples{
# Load the English NER tagger
tagger_en <- load_tagger_ner("en")

}
