% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_tagger_pos}
\alias{load_tagger_pos}
\title{Load Flair POS Tagger}
\usage{
load_tagger_pos(language)
}
\arguments{
\item{language}{A character string indicating the desired language model. If `NULL`,
the function will default to the 'pos-fast' model. Supported language models include:
"pos", "pos-fast", "upos", "upos-fast", "pos-multi", "pos-multi-fast", "ar-pos",
"de-pos", "de-pos-tweets", "da-pos", "ml-pos", "ml-upos", "pt-pos-clinical", "pos-ukrainian".}
}
\value{
A Flair POS tagger model corresponding to the specified (or default) language.
}
\description{
This function loads the POS (part-of-speech) tagger model for a specified language
using the Flair library. If no language is specified, it defaults to 'pos-fast'.
}
\examples{
\dontrun{
tagger <- load_tagger_pos("pos-fast")
}
}
