% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flaiR_loader.R
\name{load_tagger_sentiments}
\alias{load_tagger_sentiments}
\title{Load a Sentiment or Language Tagger Model from Flair}
\usage{
load_tagger_sentiments(language = NULL)
}
\arguments{
\item{language}{A character string specifying the language model to load.
Supported models include:
\itemize{
  \item "sentiment" - Sentiment analysis model
  \item "sentiment-fast" - Faster sentiment analysis model
  \item "de-offensive-language" - German offensive language detection model
} If not provided, the function will default to the "sentiment" model.}
}
\value{
An object of the loaded Flair model.
}
\description{
This function loads a pre-trained sentiment or language tagger
from the Flair library.
}
\examples{
\dontrun{
  tagger <- load_tagger_sentiments("sentiment")
}

}
