% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_flair.R
\name{prep_source}
\alias{prep_source}
\title{Helper for \code{knit_print.with_flair}}
\usage{
prep_source(x, doc_type = "unknown")
}
\arguments{
\item{x}{Text of source code.}

\item{doc_type}{Document type to knit to.}
}
\value{
Properly wrapped text.
}
\description{
Helper for \code{knit_print.with_flair}
}
